Implementing tests for the slot provider
