--- src/pages/journal/utils/newSlotsDetectingChanges.ts
@@ -1,28 +1,0 @@
-import { ExaminerWorkSchedule } from '../../../common/domain/DJournal';
-import { SlotItem } from '../../../providers/slot-selector/slot-item';
-import { DeepDiff } from 'deep-diff';
-import { flatten } from 'lodash';
-
-export default function(oldJournalSlots: SlotItem[], newJournal: ExaminerWorkSchedule): SlotItem[] {
-  const newSlots = flatten([
-    newJournal.testSlots || [],
-    newJournal.nonTestActivities || [],
-  ]);
-
-  newSlots.sort((slotA, slotB) => slotA.slotDetail.start < slotB.slotDetail.start ? -1 : 1);
-
-  return newSlots.map(newSlot => {
-    const newSlotId = newSlot.slotDetail.slotId;
-    const replacedJournalSlot = oldJournalSlots.find(oldSlot => oldSlot.slotData.slotDetail.slotId === newSlotId);
-
-    let differenceFound = false;
-    if (replacedJournalSlot) {
-      const differenceToSlot = DeepDiff(replacedJournalSlot.slotData, newSlot);
-      if (Array.isArray(differenceToSlot) && differenceToSlot.some(change => change.kind === 'E')) {
-        differenceFound = true;
-      }
-    }
-
-    return new SlotItem(newSlot, differenceFound)
-  });
-}
